+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Fixed bug when device id was changed without server merge. Previously the new session was started with the old device ID and not the new one.
 * Fixed a bug that was a typo ('COUNTLY_CUSTOM_HTTP' instead of 'COUNTLY_USE_CUSTOM_HTTP') in the cmake file that cause the SDK to be misconfigured. 
 * Fixed a bug that caused POST requests to fail on Windows.
+* Fixed issues with location requests.
+* Deprecated old location calls and introduced a new location call
 
 21.11.0
 * Fixed session duration issue.